+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/SynchronizationTask.kt
@@ -2,22 +2,15 @@ package com.flipperdevices.bridge.synchronization.impl
 
 import com.flipperdevices.bridge.api.manager.ktx.state.ConnectionState
 import com.flipperdevices.bridge.api.manager.ktx.state.FlipperSupportedState
-import com.flipperdevices.bridge.dao.api.delegates.FavoriteApi
-import com.flipperdevices.bridge.dao.api.delegates.FlipperFileApi
-import com.flipperdevices.bridge.dao.api.delegates.key.DeleteKeyApi
 import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
-import com.flipperdevices.bridge.dao.api.delegates.key.UpdateKeyApi
-import com.flipperdevices.bridge.dao.api.delegates.key.UtilsKeyApi
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
 import com.flipperdevices.bridge.service.api.FlipperServiceApi
 import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.bridge.synchronization.api.SynchronizationState
-import com.flipperdevices.bridge.synchronization.impl.executor.FlipperKeyStorage
+import com.flipperdevices.bridge.synchronization.impl.di.DaggerTaskSynchronizationComponent
 import com.flipperdevices.bridge.synchronization.impl.model.RestartSynchronizationException
-import com.flipperdevices.bridge.synchronization.impl.repository.FavoriteSynchronization
-import com.flipperdevices.bridge.synchronization.impl.repository.KeysSynchronization
-import com.flipperdevices.bridge.synchronization.impl.repository.storage.ManifestRepository
-import com.flipperdevices.bridge.synchronization.impl.utils.SynchronizationPercentProvider
+import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.core.di.ComponentHolder
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
@@ -25,27 +18,33 @@ import com.flipperdevices.core.ui.lifecycle.OneTimeExecutionBleTask
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.complex.SynchronizationEnd
 import com.flipperdevices.wearable.sync.handheld.api.SyncWearableApi
+import com.squareup.anvil.annotations.ContributesBinding
+import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.withContext
 
-@Suppress("LongParameterList")
-class SynchronizationTask(
-    private val serviceProvider: FlipperServiceProvider,
-    private val simpleKeyApi: SimpleKeyApi,
-    private val deleteKeyApi: DeleteKeyApi,
-    private val utilsKeyApi: UtilsKeyApi,
-    private val favoriteApi: FavoriteApi,
-    private val metricApi: MetricApi,
-    private val synchronizationProvider: SynchronizationPercentProvider,
-    private val flipperFileApi: FlipperFileApi,
-    private val syncWearableApi: SyncWearableApi,
-    private val updateKeyApi: UpdateKeyApi
-) : OneTimeExecutionBleTask<Unit, SynchronizationState>(serviceProvider), LogTagProvider {
-    override val TAG = "SynchronizationTask"
+interface SynchronizationTask {
+    fun start(
+        input: Unit,
+        stateListener: suspend (SynchronizationState) -> Unit
+    ): Job
 
-    private val manifestRepository = ManifestRepository()
+    suspend fun onStop()
+}
+
+@ContributesBinding(AppGraph::class, SynchronizationTask::class)
+class SynchronizationTaskImpl @Inject constructor(
+    serviceProvider: FlipperServiceProvider,
+    private val simpleKeyApi: SimpleKeyApi,
+    private val metricApi: MetricApi,
+    private val syncWearableApi: SyncWearableApi
+) : OneTimeExecutionBleTask<Unit, SynchronizationState>(serviceProvider),
+    SynchronizationTask,
+    LogTagProvider {
+    override val TAG = "SynchronizationTask"
 
     override suspend fun startInternal(
         scope: CoroutineScope,
@@ -94,30 +93,19 @@ class SynchronizationTask(
         serviceApi: FlipperServiceApi,
         onStateUpdate: suspend (SynchronizationState) -> Unit
     ) = withContext(Dispatchers.Default) {
-        val flipperStorage = FlipperKeyStorage(serviceApi.requestApi)
-        val favoriteSynchronization = FavoriteSynchronization(
-            favoriteApi,
-            manifestRepository,
-            flipperStorage
-        )
-        val keysSynchronization = KeysSynchronization(
-            simpleKeyApi,
-            deleteKeyApi,
-            utilsKeyApi,
-            manifestRepository,
-            flipperStorage,
-            serviceApi.requestApi,
-            synchronizationProvider,
-            flipperFileApi,
-            updateKeyApi
-        )
+        val taskComponent = DaggerTaskSynchronizationComponent.factory()
+            .create(
+                ComponentHolder.component(),
+                serviceApi.requestApi
+            )
 
-        val keysHashes = keysSynchronization.syncKeys(onStateUpdate)
-        val favorites = favoriteSynchronization.syncFavorites()
+        val keysHashes = taskComponent.keysSynchronization.syncKeys(onStateUpdate)
+        val favorites = taskComponent.favoriteSynchronization.syncFavorites()
 
         // End synchronization keys
-        manifestRepository.saveManifest(keysHashes, favorites)
-        synchronizationProvider.markedAsFinish()
+        taskComponent.manifestRepository.saveManifest(keysHashes, favorites)
+        taskComponent.synchronizationProvider.markedAsFinish()
+
         try {
             syncWearableApi.updateWearableIndex()
         } catch (throwable: Exception) {