+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/api/SynchronizationApiImpl.kt
@@ -1,23 +1,11 @@
 package com.flipperdevices.bridge.synchronization.impl.api
 
-import androidx.datastore.core.DataStore
-import com.flipperdevices.bridge.dao.api.delegates.FavoriteApi
-import com.flipperdevices.bridge.dao.api.delegates.FlipperFileApi
-import com.flipperdevices.bridge.dao.api.delegates.key.DeleteKeyApi
-import com.flipperdevices.bridge.dao.api.delegates.key.SimpleKeyApi
-import com.flipperdevices.bridge.dao.api.delegates.key.UpdateKeyApi
-import com.flipperdevices.bridge.dao.api.delegates.key.UtilsKeyApi
-import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
 import com.flipperdevices.bridge.synchronization.api.SynchronizationApi
 import com.flipperdevices.bridge.synchronization.api.SynchronizationState
 import com.flipperdevices.bridge.synchronization.impl.SynchronizationTask
-import com.flipperdevices.bridge.synchronization.impl.utils.SynchronizationPercentProvider
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
-import com.flipperdevices.core.preference.pb.Settings
-import com.flipperdevices.metric.api.MetricApi
-import com.flipperdevices.wearable.sync.handheld.api.SyncWearableApi
 import com.squareup.anvil.annotations.ContributesBinding
 import java.util.concurrent.atomic.AtomicBoolean
 import javax.inject.Inject
@@ -30,16 +18,7 @@ import kotlinx.coroutines.flow.update
 @ContributesBinding(AppGraph::class, SynchronizationApi::class)
 @Suppress("LongParameterList")
 class SynchronizationApiImpl @Inject constructor(
-    private val serviceProvider: FlipperServiceProvider,
-    private val simpleKeyApi: SimpleKeyApi,
-    private val deleteKeyApi: DeleteKeyApi,
-    private val utilsKeyApi: UtilsKeyApi,
-    private val favoriteApi: FavoriteApi,
-    private val metricApi: MetricApi,
-    private val preference: DataStore<Settings>,
-    private val flipperFileApi: FlipperFileApi,
-    private val syncWearableApi: SyncWearableApi,
-    private val updateKeyApi: UpdateKeyApi
+    private val synchronizationTask: SynchronizationTask
 ) : SynchronizationApi, LogTagProvider {
     override val TAG = "SynchronizationApi-${hashCode()}"
 
@@ -48,8 +27,6 @@ class SynchronizationApiImpl @Inject constructor(
         SynchronizationState.NotStarted
     )
     private var markDirty = false
-
-    private var synchronizationTask: SynchronizationTask? = null
 
     @Synchronized
     override fun startSynchronization(force: Boolean) {
@@ -63,24 +40,12 @@ class SynchronizationApiImpl @Inject constructor(
         }
         markDirty = false
         synchronizationState.update { SynchronizationState.InProgress(0f) }
-        val localSynchronizationTask = SynchronizationTask(
-            serviceProvider = serviceProvider,
-            simpleKeyApi = simpleKeyApi,
-            deleteKeyApi = deleteKeyApi,
-            utilsKeyApi = utilsKeyApi,
-            favoriteApi = favoriteApi,
-            metricApi = metricApi,
-            synchronizationProvider = SynchronizationPercentProvider(preference),
-            flipperFileApi = flipperFileApi,
-            syncWearableApi = syncWearableApi,
-            updateKeyApi = updateKeyApi
-        )
-        info { "Create synchronization task ${localSynchronizationTask.hashCode()}" }
+        info { "Create synchronization task ${synchronizationTask.hashCode()}" }
 
-        localSynchronizationTask.start(input = Unit) { taskState ->
+        synchronizationTask.start(input = Unit) { taskState ->
             synchronizationState.update { taskState }
             if (taskState == SynchronizationState.Finished) {
-                localSynchronizationTask.onStop()
+                synchronizationTask.onStop()
                 isLaunched.compareAndSet(true, false)
                 info { "Mark launched false" }
                 if (markDirty) {
@@ -88,12 +53,11 @@ class SynchronizationApiImpl @Inject constructor(
                 }
             }
         }
-        synchronizationTask = localSynchronizationTask
     }
 
     override suspend fun stop() {
         markDirty = false
-        synchronizationTask?.onStop()
+        synchronizationTask.onStop()
     }
 
     override fun isSynchronizationRunning() = isLaunched.get()