+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/di/TaskGraph.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.bridge.synchronization.impl.di
+
+abstract class TaskGraph private constructor()