+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/storage/ManifestStorage.kt
@@ -2,19 +2,34 @@ package com.flipperdevices.bridge.synchronization.impl.repository.storage
 
 import android.content.Context
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
+import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
 import com.flipperdevices.bridge.synchronization.impl.model.KeyWithHash
 import com.flipperdevices.bridge.synchronization.impl.model.ManifestFile
+import com.flipperdevices.core.di.SingleIn
+import com.squareup.anvil.annotations.ContributesBinding
 import java.io.File
+import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromStream
 import kotlinx.serialization.json.encodeToStream
 
-class ManifestStorage(context: Context) {
+interface ManifestStorage {
+    suspend fun save(
+        keys: List<KeyWithHash>,
+        favorites: List<FlipperFilePath>
+    )
+
+    suspend fun load(): ManifestFile?
+}
+
+@SingleIn(TaskGraph::class)
+@ContributesBinding(TaskGraph::class)
+class ManifestStorageImpl @Inject constructor(context: Context) : ManifestStorage {
     private val file: File = File(context.filesDir, "LastSyncManifest_v4.json")
 
-    suspend fun save(
+    override suspend fun save(
         keys: List<KeyWithHash>,
         favorites: List<FlipperFilePath>
     ) = withContext(Dispatchers.IO) {
@@ -26,7 +41,7 @@ class ManifestStorage(context: Context) {
         }
     }
 
-    suspend fun load(): ManifestFile? = withContext(Dispatchers.IO) {
+    override suspend fun load(): ManifestFile? = withContext(Dispatchers.IO) {
         if (!file.exists()) {
             return@withContext null
         }