+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [REFACTOR] Display error if /int/ full in update progress
 - [REFACTOR] Markdown compose bump
 - [REFACTOR] New ci with upload wearos app
+- [REFACTOR] Migrate synchronization task graph to TaskSynchronizationComponent
 - [Feature] Add support for NFC Shadow Files
 - [Feature] Offer update, if region changes or region file not exit on Flipper
 - [Feature] Do not provide SubGhz provisioning on zero hardware region